requirements.txt
tqdm
Pillow
torch==1.13.1
torchvision==0.14.1
numpy==1.26.4
gradio==4.27.0
einops==0.7.0
omegaconf==2.3.0
imageio==2.34.1
imageio-ffmpeg==0.4.9
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
pytorch-lightning==1.5.0
torchmetrics==1.3.2
accelerate==0.29.3
transformers==4.19.2
timm==0.9.16
open_clip_torch==2.24.0
lpips==0.1.4
basicsr==1.4.2
pycocotools==2.0.7
invisible-watermark==0.1.5
kornia==0.7.2
scikit-learn==1.4.2
